fix: 잘못된 layout 적용으로 일부 페이지에 tailwind css가 적용되지 않는 오류 해결
