+++ src/app/(daum)/layout.tsx
@@ -1,28 +1,14 @@
-import type { Metadata } from 'next'
 import MainBottomNavigationBar from '@/shared/components/MainBottomNavigationBar'
-import '@/shared/styles/globals.css'
-import '@/shared/styles/fonts.css'
 
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
-
-export default function RootLayout({
+export default function DaumLayout({
   children,
 }: Readonly<{
   children: React.ReactNode
 }>) {
   return (
-    <html lang="ko">
-      <body>
-        <div className="flex justify-center">
-          <main className="flex h-svh w-full max-w-md flex-col">
-            <section className="flex-1 overflow-auto">{children}</section>
-            <MainBottomNavigationBar />
-          </main>
-        </div>
-      </body>
-    </html>
+    <div className="flex h-svh w-full max-w-md flex-col">
+      <div className="flex-1 overflow-auto">{children}</div>
+      <MainBottomNavigationBar />
+    </div>
   )
 }
