+++ src/app/layout.tsx
@@ -1,0 +1,22 @@
+import type { Metadata } from 'next'
+import '@/shared/styles/globals.css'
+import '@/shared/styles/fonts.css'
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode
+}>) {
+  return (
+    <html lang="ko">
+      <body>
+        <div className="flex justify-center">{children}</div>
+      </body>
+    </html>
+  )
+}
